Fixed issue with Medium doing 'soft reloads'
Medium uses 'soft page loads' where the content of the page is 80% the
same and only changes completely when a new hostname is entered. This
was causing issues with bookmarklets not showing up and as such, a
method for fixing this was required.

Ended up wathcing the url and hostname and when the url changes, then
force a hard page load which 1) causes the bookmarklet to appear
always and 2) just looks like the page loaded normally to most users.

This commit shows the actual TODO being checked off - the above code
changes are found within the `startScript()` function in `bookmarklets-
.js`.