Updating of the domain list via Settings UI
This brings a functioning settings ui which allows for adding and remo-
ving of domains to the gloabl domain list. The 'Add Domain' button cre-
ates a new domain component which contains a text input element and a
button for removing that component. Upon saving, the values from the i-
nput fields are fetched then set to local storage under the global
domain list (used to allow/disallow medium-bookmarklet functionality f-
or a given domain).