Fixed empty state appearing while settings active
Previously there was an issue where adding or removing a bookmarklet
while the settings page was active would cause the bookmarklets to
appear. This was fixed by separating the settings and main content divs
into two. It was discovered that when removing a card with the settings
page active, the empty state container would show up if there were no
cards. This has been fixed by putting the empty state container inside
the main content div.