Notification displayed upon settings saved (#4)
This commit adds a notification that drops down when the settings have
been saved. After two seconds, the notification automatically closes.
This is a partial fix for #4. To close, will need to add a notification
when bookmarklets are added or deleted in Medium.com itself (via content
script - `bookmarklets.js`).