Implement overflow menu using Buefy dropdown
Using a Buefy dropdown component, was able to replicate the design for the
overflow menu. Current menu items include: Domain List, Highlight Color, and
Dark Mode toggle, plus a static (and hardcoded) version number.

This works decently with the smallest viewport of the sidebar, tweaks will be
needed however. So far, good progress!